refactor: Navigationをリファクタ
